Move Sysconfig*File classes to Sysconfig module
